fix post body rendering
